Extract AddEditSlackSidebar into file
