+++ frontend/src/metabase/sharing/components/AddEditSidebar/AddEditSidebar.jsx
@@ -1,27 +1,7 @@
-/* eslint "react/prop-types": "error" */
-
-import React from "react";
-import PropTypes from "prop-types";
-import _ from "underscore";
-import { t } from "ttag";
 import { connect } from "react-redux";
-import { CHANNEL_NOUN_PLURAL } from "./constants";
 
 import _AddEditEmailSidebar from "./AddEditEmailSidebar";
-import ChannelFields from "./ChannelFields";
-import DefaultParametersSection from "./DefaultParametersSection";
-import DeleteSubscriptionAction from "./DeleteSubscriptionAction";
-import Heading from "./Heading";
-import CaveatMessage from "./CaveatMessage";
-import Icon from "metabase/components/Icon";
-import SchedulePicker from "metabase/components/SchedulePicker";
-import SendTestPulse from "metabase/components/SendTestPulse";
-import Sidebar from "metabase/dashboard/components/Sidebar";
-import Toggle from "metabase/core/components/Toggle";
-
-import { dashboardPulseIsValid } from "metabase/lib/pulse";
-
-import { PLUGIN_DASHBOARD_SUBSCRIPTION_PARAMETERS_SECTION_OVERRIDE } from "metabase/plugins";
+import _AddEditSlackSidebar from "./AddEditSlackSidebar";
 
 import {
   getDefaultParametersById,
@@ -41,123 +21,3 @@ export const AddEditEmailSidebar = connect(mapStateToProps)(
 export const AddEditSlackSidebar = connect(mapStateToProps)(
   _AddEditSlackSidebar,
 );
-
-function _AddEditSlackSidebar({
-  pulse,
-  formInput,
-  channel,
-  channelSpec,
-  parameters,
-  defaultParametersById,
-  dashboard,
-  // form callbacks
-  handleSave,
-  onCancel,
-  onChannelPropertyChange,
-  onChannelScheduleChange,
-  testPulse,
-  toggleSkipIfEmpty,
-  handleArchive,
-  setPulseParameters,
-}) {
-  const isValid = dashboardPulseIsValid(pulse, formInput.channels);
-
-  return (
-    <Sidebar
-      closeIsDisabled={!isValid}
-      onClose={handleSave}
-      onCancel={onCancel}
-    >
-      <div className="pt4 flex align-center px4 mb3">
-        <Icon name="slack" className="mr1" size={21} />
-        <Heading>{t`Send this dashboard to Slack`}</Heading>
-      </div>
-      <CaveatMessage />
-      <div className="pb2 px4">
-        {channelSpec.fields && (
-          <ChannelFields
-            channel={channel}
-            channelSpec={channelSpec}
-            onChannelPropertyChange={onChannelPropertyChange}
-          />
-        )}
-        <SchedulePicker
-          schedule={_.pick(
-            channel,
-            "schedule_day",
-            "schedule_frame",
-            "schedule_hour",
-            "schedule_type",
-          )}
-          scheduleOptions={channelSpec.schedules}
-          textBeforeInterval={t`Sent`}
-          textBeforeSendTime={t`${CHANNEL_NOUN_PLURAL[
-            channelSpec && channelSpec.type
-          ] || t`Messages`} will be sent at`}
-          onScheduleChange={(newSchedule, changedProp) =>
-            onChannelScheduleChange(newSchedule, changedProp)
-          }
-        />
-        <div className="pt2 pb1">
-          <SendTestPulse
-            channel={channel}
-            channelSpecs={formInput.channels}
-            pulse={pulse}
-            testPulse={testPulse}
-            normalText={t`Send to Slack now`}
-            successText={t`Slack sent`}
-            disabled={!isValid}
-          />
-        </div>
-        {PLUGIN_DASHBOARD_SUBSCRIPTION_PARAMETERS_SECTION_OVERRIDE.Component ? (
-          <PLUGIN_DASHBOARD_SUBSCRIPTION_PARAMETERS_SECTION_OVERRIDE.Component
-            className="py3 mt2 border-top"
-            parameters={parameters}
-            dashboard={dashboard}
-            pulse={pulse}
-            setPulseParameters={setPulseParameters}
-            defaultParametersById={defaultParametersById}
-          />
-        ) : (
-          <DefaultParametersSection
-            className="py3 mt2 border-top"
-            parameters={parameters}
-            defaultParametersById={defaultParametersById}
-          />
-        )}
-        <div className="text-bold py2 flex justify-between align-center border-top">
-          <Heading>{t`Don't send if there aren't results`}</Heading>
-          <Toggle
-            value={pulse.skip_if_empty || false}
-            onChange={toggleSkipIfEmpty}
-          />
-        </div>
-        {pulse.id != null && (
-          <DeleteSubscriptionAction
-            pulse={pulse}
-            handleArchive={handleArchive}
-          />
-        )}
-      </div>
-    </Sidebar>
-  );
-}
-
-_AddEditSlackSidebar.propTypes = {
-  pulse: PropTypes.object.isRequired,
-  formInput: PropTypes.object.isRequired,
-  channel: PropTypes.object.isRequired,
-  channelSpec: PropTypes.object.isRequired,
-  users: PropTypes.array,
-  parameters: PropTypes.array.isRequired,
-  defaultParametersById: PropTypes.object.isRequired,
-  dashboard: PropTypes.object.isRequired,
-  handleSave: PropTypes.func.isRequired,
-  onCancel: PropTypes.func.isRequired,
-  onChannelPropertyChange: PropTypes.func.isRequired,
-  onChannelScheduleChange: PropTypes.func.isRequired,
-  testPulse: PropTypes.func.isRequired,
-  toggleSkipIfEmpty: PropTypes.func.isRequired,
-  handleArchive: PropTypes.func.isRequired,
-  setPulseParameters: PropTypes.func.isRequired,
-};
